add additional 'cluster' and 'dist_to_center' feature for FS data
In the SADIT software, when considering data format with IP addresses. We add tow additional feature.  One is cluster label of the src_ip address, the other is the distance to the cluster center of the src_ip address. In the early stage of GAD, these two features are removed to make the method general. This commit add the two features to the GAD by adding a new Data file.

 It also improves the structure. and fix a bug in get_min_max() funciton.