Rename load_para as load_config. Make it support json file.
Previously the config file has to be a python exec file. This CL improve
load_config so that it can also load json config file. The file type is
justed based on suffix. File ending with '.py' will be executed as python
file and file ending with '.json' will be loaded as json file.