*Allow user to specify date_parser in config file to increase CSVFile parse speed.
Previously, the CSVFile parse is very slow become the default behavior
of pandas.read_csv needs to detect the format of date for every line. In
this CL, we add a new configuration option date_parser to allow user to
specify date_parser.

date_parser can either be a format string or a function. If it is a
string, a parser function will be contructed using
pandas.datetime.strptime.

TESTED=sample run. For the following command
./cmdgad  timebased_botnet_detection_eval -c
./example-configs/detect-config-botnet.py -d
../dataset/CTU-Malware-Capture-Botnet-47/detailed-bidirectional-flow-labels/capture20110816.binetflow
The speed comparison is:
method: time
no option: 1m48.536s
date_parser as string: 21s
date_parser as fast_date_parser func: 4.2s